Add Pre-Alpha build code for ultrasonic distance sensor
Code for tinkercad pre-alpha build prototype. The parts used here will not be used in the future and we will most likely use a raspberry pi. This is purely meant to demonstrates the core of our project, motion/distance detection that then relays haptic feedback to alert the biker of potential dangers.